refactor: move constraint parsing of generic constraints from __init__ into parser module and rename module for constraint parsing of version constraints from helpers to parser
