+++ src/poetry/core/packages/constraints/__init__.py
@@ -2,6 +2,15 @@ from __future__ import annotations
 
 import warnings
 
+from poetry.core.constraints.generic import AnyConstraint
+from poetry.core.constraints.generic import BaseConstraint
+from poetry.core.constraints.generic import Constraint
+from poetry.core.constraints.generic import EmptyConstraint
+from poetry.core.constraints.generic import MultiConstraint
+from poetry.core.constraints.generic import UnionConstraint
+from poetry.core.constraints.generic import parse_constraint
+from poetry.core.constraints.generic.parser import parse_single_constraint
+
 
 warnings.warn(
     "poetry.core.packages.constraints is deprecated."
@@ -10,4 +19,14 @@ warnings.warn(
     stacklevel=2,
 )
 
-from poetry.core.constraints.generic import *  # noqa: E402, F401, F403
+
+__all__ = [
+    "AnyConstraint",
+    "BaseConstraint",
+    "Constraint",
+    "EmptyConstraint",
+    "MultiConstraint",
+    "UnionConstraint",
+    "parse_constraint",
+    "parse_single_constraint",
+]
