+++ src/poetry/core/semver/helpers.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from poetry.core.constraints.version import parse_constraint
+from poetry.core.constraints.version.parser import parse_constraint
+from poetry.core.constraints.version.parser import parse_single_constraint
 
 
-__all__ = ["parse_constraint"]
+__all__ = ["parse_constraint", "parse_single_constraint"]
